Upgrade to latest node-cpp-skel build scripts

+++ src/matrix.h
@@ -16,14 +16,14 @@ class Matrix {
         if (n < 0)
             throw std::runtime_error{"Negative dimension"};
 
-        data.resize(n * n);
+        data.resize(static_cast<std::size_t>(n * n));
     }
 
     std::int32_t dim() const { return n; }
     std::int32_t size() const { return dim() * dim(); }
 
-    T& at(std::int32_t x, std::int32_t y) { return data.at(y * n + x); }
-    const T& at(std::int32_t x, std::int32_t y) const { return data.at(y * n + x); }
+    T& at(std::int32_t x, std::int32_t y) { return data.at(static_cast<std::size_t>(y * n + x)); }
+    const T& at(std::int32_t x, std::int32_t y) const { return data.at(static_cast<std::size_t>(y * n + x)); }
 
   private:
     std::int32_t n;
