+++ src/vector.h
@@ -10,12 +10,12 @@ class Vector {
     using Value = T;
 
     Vector() = default;
-    Vector(std::int32_t n) { data.resize(n); }
+    Vector(std::int32_t n) { data.resize(static_cast<std::size_t>(n)); }
 
-    std::int32_t size() const { return data.size(); }
+    std::int32_t size() const { return static_cast<std::int32_t>(data.size()); }
 
-    T& at(std::int32_t x) { return data.at(x); }
-    const T& at(std::int32_t x) const { return data.at(x); }
+    T& at(std::int32_t x) { return data.at(static_cast<std::size_t>(x)); }
+    const T& at(std::int32_t x) const { return data.at(static_cast<std::size_t>(x)); }
 
   private:
     std::vector<T> data;
